requirements.txt
python-telegram-bot==21.6
Flask==3.0.3
requests==2.31.0